decision: updated "deferred plugin calls" decision & implementation
